Modularize: Make ES build emit separately-importable files
