+++ size-demo/getLeyningForParsha.ts
@@ -1,2 +1,2 @@
-import {getLeyningForParsha} from '../dist';
+import {getLeyningForParsha} from '../dist/es/leyning';
 console.log(getLeyningForParsha('Noach'));
