+++ size-demo/sizes.md
@@ -1,6 +1,6 @@
 # Bundle sizes
 
-- `getLeyningForParsha.mjs`: 69.2 kB
-- `getLeyningForParsha.min.mjs`: 45.1 kB
-- `getLeyningOnDate.mjs`: 205 kB
-- `getLeyningOnDate.min.mjs`: 93.9 kB
+- `getLeyningForParsha.mjs`: 65.7 kB
+- `getLeyningForParsha.min.mjs`: 41.9 kB
+- `getLeyningOnDate.mjs`: 202 kB
+- `getLeyningOnDate.min.mjs`: 90.9 kB
